[FEAT] email 전송 스케줄러 구현

+++ api/src/main/kotlin/boaz/lol/co/application/subscribe/service/SubscribeService.kt
@@ -11,35 +11,29 @@ class SubscribeService(
     private val subscribeRepository: SubscribeRepository,
     private val accountRepository: AccountRepository,
     private val championRepository: ChampionRepository,
-    private val emailService: EmailService
 ) {
 
-    fun addSubscription(req: SubscribeReq) {
-        val account = accountRepository.getById(req.accountId).orElseThrow()
+
+    fun like(req: SubscribeReq, accountData: AccountData) {
+        val account = accountRepository.getById(accountData.id).orElseThrow()
         val champion = championRepository.getById(req.championId).orElseThrow()
-        subscribeRepository.addSubscribe(account, champion)
-    }
 
-    fun removeSubscription(req: SubscribeReq) {
-        val account = accountRepository.getById(req.accountId).orElseThrow()
-        val champion = championRepository.getById(req.championId).orElseThrow()
-        subscribeRepository.deleteSubscribe(account, champion)
-    }
-
-    fun getSubscribeData(accountId: Long): SubscribeData {
-        val account = accountRepository.getById(accountId).orElseThrow()
-        val subscribe = subscribeRepository.getByAccount(account)
-        return subscribe.toSubscribeData()
-    }
-
-    fun notify(champions: List<Champion>) {
-        val accounts = subscribeRepository.getAccountsByChampions(champions)
-        accounts.forEach { account ->
-            emailService.sendChampionUpdateNotification(
-                account.email,
-                champions.map { it.name }.joinToString(", "),
-                "업데이트 챔피언: ${champions.joinToString { it.name }}"
-            )
+        if (req.subscribe) {
+            subscribeRepository.addSubscribe(account, champion)
+        } else {
+            subscribeRepository.deleteSubscribe(account, champion)
         }
     }
+
+
+//    fun notify(champions: List<Champion>) {
+//        val accounts = subscribeRepository.getAccountsByChampions(champions)
+//        accounts.forEach { account ->
+//            emailService.sendChampionUpdateNotification(
+//                account.email,
+//                champions.map { it.name }.joinToString(", "),
+//                "업데이트 챔피언: ${champions.joinToString { it.name }}"
+//            )
+//        }
+//    }
 }
